+++ ReactiveCocoa/UIKit/iOS/UIKeyboard.swift
@@ -19,7 +19,8 @@ public struct KeyboardChangeContext {
 	/// The animation curve which the system keyboard will use to animate the
 	/// change in its frame.
 	public var animationCurve: UIViewAnimationCurve {
-		return base[UIKeyboardAnimationCurveUserInfoKey] as! UIViewAnimationCurve
+		let value = base[UIKeyboardAnimationCurveUserInfoKey] as! NSNumber
+		return UIViewAnimationCurve(rawValue: value.intValue)!
 	}
 
 	/// The duration in which the system keyboard expects to animate the change in